Backend/requirements.txt
fastapi
uvicorn
requests
httpx
pytest
python-dotenv
google-generativeai